Store ref counts as unions.
This lets us tag the ref count more accurately and lessen the number of deallocations internally in the engine. We can store the initial ref count and know when we need to actually send it back to the engine for dereferencing, otherwise we can just decrement the reference counter in Intercept.